fix: storage adapter migration + crypto utils + PDF service fix (dart analyze 0 errors)
- Move storage_adapter.dart from editor_core to main app (cross-package dependency fix)
- Add sha256Hex, deriveKeyFromPassword, secureRandomBytes to crypto_utils.dart (PointCastle)
- Replace Printing.raster with pdfrx in unified_pdf_service.dart
- Fix envelope_encryption/schema_version/three_layer_encryption unused imports and casts
- Add schema_migration_crypto_test.dart: 19 tests (SchemaMigration/sha256Hex/deriveKey/secureRandom/DataIntegrity)
- Export schema_version.dart and unified_storage.dart from editor_core barrel
- Add path dependency to pubspec.yaml

+++ packages/editor_core/lib/src/domain/three_layer_encryption.dart
@@ -633,7 +633,7 @@ class ThreeLayerEncryptionService {
   /// 生成随机填充长度（16~256 字节——CSPRNG——隐藏真实大小）。
   int _randomPaddingLength() {
     final secureRandom = _createSecureRandom();
-    return 16 + (secureRandom.nextUint8() as int) % 241; // 16 + [0..240] = [16..256]。
+    return 16 + secureRandom.nextUint8() % 241; // 16 + [0..240] = [16..256]。
   }
 
   /// 构造填充字节：[random:Nbytes][length:2bytes]——长度在末尾（方便解密时直接读取）。
@@ -641,7 +641,7 @@ class ThreeLayerEncryptionService {
     final secureRandom = _createSecureRandom();
     final result = Uint8List(length + 2);
     for (var i = 0; i < length; i++) {
-      result[i] = secureRandom.nextUint8() as int;
+      result[i] = secureRandom.nextUint8();
     }
     result[length] = (length >> 8) & 0xFF;
     result[length + 1] = length & 0xFF;
@@ -653,7 +653,7 @@ class ThreeLayerEncryptionService {
     final secureRandom = _createSecureRandom();
     final nonce = Uint8List(length);
     for (var i = 0; i < length; i++) {
-      nonce[i] = secureRandom.nextUint8() as int;
+      nonce[i] = secureRandom.nextUint8();
     }
     return nonce;
   }
